pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pamssw"
version = "0.1.0"
description = "Low-parameter SSW and LS-SSW global optimization toolkit"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "PyYAML>=6.0",
  "ase>=3.22",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
pamssw = "pamssw.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
